Revert "Revert "Merge pull request #44 from gpsubprojectfirst/hy""
This reverts commit 71e11ce527cf22683b1741b57a66fe883cb3bc93.

+++ 0806_임시프로젝트파일_ 하영/GP_subproject_first/myUnit.h
@@ -30,7 +30,9 @@ public:
 	virtual void Rest(float Delta) {}
 	virtual void ExtraAction(float Delta);
 	virtual void CreateBT(BlackBoard* InBB);
-	virtual void CalcDirection(int xvec,int yvec);
+
+	virtual void CalcDirection(float xvec,float yvec);
+
 
 	//unit status
 	UnitInfo mUnitInfo;
